Split out copy_op from utility_ops (#11470)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/11470

In order to reduce build sizes, we are identifying files that can be split up into smaller units, allowing us to only include the ops we need.

Reviewed By: ajtulloch

Differential Revision: D9725819

fbshipit-source-id: 95219da2d645c56200f9a554fba47cfbf6471b1c
